requirements.txt
streamlit==1.32.2
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.1
